+++ CHANGELOG.md
@@ -4,7 +4,14 @@ This project adheres to [Semantic Versioning 2.0](http://semver.org/) (semver). 
 currently in 0.y.x development and there may be breaking changes between minor releases, 
 though we strive to limit breaking changes as much as possible.
 
-## Unreleased
+## v0.5.0 - 2017-10-27
+
+#### Added
+
+- Added a new 'Your Keys' tab that allows you to track what keys you have across all 
+of your characters. *Please note that the key will not appear until you have logged onto 
+that character each week.*
+- Added a minimap button to show your keystones as a tooltip and to open the main ledger window when clicked on. You may turn the button off with /mplus button off and turn it back on with /mplus button on.
 
 #### Changed